third_party/blink/web_tests/external/wpt/css/css-shadow/font-face-005-expected.txt

This is a testharness.js-based test.
[FAIL] @font-face should not leak out of shadow tree.
  assert_equals: expected 160 but got 80
Harness: the test ran to completion.
